@@ -5,28 +5,36 @@ Revises: 20250617_add_book_status
 Create Date: 2025-06-18 23:01:09.599133
 
 """
-from typing import Sequence, Union
 
-from alembic import op
+from collections.abc import Sequence
+
 import sqlalchemy as sa
 
+from alembic import op
 
 # revision identifiers, used by Alembic.
-revision: str = '6ce246100fed'
-down_revision: Union[str, None] = '20250617_add_book_status'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
+revision: str = "6ce246100fed"
+down_revision: str | None = "20250617_add_book_status"
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
 
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('books', sa.Column('user_id', sa.String(), nullable=False))
-    op.create_foreign_key(None, 'books', 'users', ['user_id'], ['id'], ondelete='CASCADE')
+    op.add_column("books", sa.Column("user_id", sa.String(), nullable=False))
+    op.create_foreign_key(
+        None,
+        "books",
+        "users",
+        ["user_id"],
+        ["id"],
+        ondelete="CASCADE",
+    )
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'books', type_='foreignkey')
-    op.drop_column('books', 'user_id')
+    op.drop_constraint(None, "books", type_="foreignkey")
+    op.drop_column("books", "user_id")
     # ### end Alembic commands ###
