@@ -1,8 +1,7 @@
 import logging
-from typing import Optional
 
 
-def get_logger(name: Optional[str] = None) -> logging.Logger:
+def get_logger(name: str | None = None) -> logging.Logger:
     logger = logging.getLogger(name or "shelf")
 
     if not logger.hasHandlers():
